Notebook Provider new functions and split
